Better caching of package metadata (unfinished)
This is the first shot at package metadata caching.
The idea is that for each remote repo, we keep a
"shadow" repo locally. The shadow repo only has the
meta-information, i.e. the PACKAGES* files, and we
also keep the Etags for the PACKAGES.gz or PACKAGES
file, to be able to decide if we need to update the
file.

The problem with this is that downloading the PACKAGES*
file is not the only bottleneck in available.packages.
The other bottlenecks are:
* reading the file with read.dcf(). available.packages()
  uses an RDS cache in the session temporary directory
  for remote repos, but not for private ones. So it would
  parse the PACKAGES file over and over again.
* Running the filters on the file. These filters
  remove duplicate packages, packages that require newer
  R versions than the current, etc. Typically the same
  set of filters run every time, so we could actually
  cache the filtered data as well.

Here is that plan for the next (proper) implementation:
* in the cache we keep the filtered RDS files, using the
  default set of filters, separately for each calling R
  version.
* We also keep the raw PACKAGES data, in case we need to
  filter it for a different R version, or a different
  set of filters in general.
* In the available.packages() shim, we are just putting
  the "right" RDS file in the R session temporary
  directory, and we call `utils::available.packages()`
  with `filters=list()`.