Cache package information, towards #14
available_packages() now uses the cached data,
but we need the other functions to do the same.
We probably need to just provide the 'available'
argument to install.packages(), etc.

Just calling available_packages() is good, too, because
this creates the cached RDS files for install.packages()
to use, but the filtering is still happening.

In the type = "both" case the filtering cannot be avoided
currently, because in this case the 'available' argument
is not allowed.